Improved communication between window + emulator
Implemented two-way message passing between the main thread and the emulator thread: A Sender and Receiver are passed to the HydraApp and Emulator, respectively, enabling commands to be sent to the emulator via the menu bar. Communication the other way is achieved through the EventLoop's built-in UserEvent proxy functionality, and is now used to render graphics on the main thread.

In addition, the Memory of the Game Boy emulator has been changed to an Rc<RefCell<>> to better accommodate sharing between components in the future.